rename app-loader to just loader
